refactor(Icon): updates styles to new configuration
